+++ protected/humhub/modules/comment/activities/views/mail/newComment.php
@@ -1,7 +1,7 @@
 <?php
 
+use humhub\modules\content\widgets\richtext\converter\RichTextToHtmlConverter;
 use humhub\modules\content\widgets\richtext\converter\RichTextToShortTextConverter;
-use humhub\modules\content\widgets\richtext\RichText;
 use yii\helpers\Html;
 
 /* @var $originator \humhub\modules\user\models\User */
@@ -13,6 +13,6 @@ echo Yii::t('CommentModule.base', "{displayName} wrote a new comment ", [
 ?>
 <br>
 
-"<?= RichText::preview($source->message, [
+"<?= RichTextToHtmlConverter::process($source->message, [
     RichTextToShortTextConverter::OPTION_CACHE_KEY => RichTextToShortTextConverter::buildCacheKeyForRecord($source)
 ]) ?>"