+++ CHANGELOG.md
@@ -13,6 +13,7 @@ HumHub Changelog
 - Fix #4985: Fix Activity Mail QueryParams on console mode
 - Fix #4989: Translate profile field title in admin list
 - Fix #5002: Fix loading of fixture spaces on tests
+- Fix #4947: Fix preview text in html mail summary
 
 
 1.8.1 (March 12, 2021)